+++ src/about/AboutPage.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import HowItWorks from './components/HowItWorks';
+import NumberedList from '../shared/components/NumberedList';
 import Contact from './components/Contact';
 import classNames from './styles/AboutPageStylesheet.css';
 
@@ -8,7 +8,23 @@ const AboutPage = () => <div className={classNames.this}>
   <p className={classNames.text}>
     This is a web platform for teaming up with people who have complementary resources and objectives.
   </p>
-  <HowItWorks />
+  <NumberedList
+    title={'How It Works'}
+    listItems={[
+      'Post resources: land, labor, equipment, and materials.',
+      'Create tasks and organizations.',
+      'Allocate resources to tasks and organizations.',
+    ]}
+  />
+  <NumberedList
+    title={'Resource Allocation'}
+    listItems={[
+      'Resource owners offer resources to tasks and organizations.',
+      'Task authors and organization members request resources.',
+      'Offers and requests have text inputs for contact info which is only shared with the person on the other end.',
+      'Users establish communication however they want, then review and decide on the proposal.',
+    ]}
+  />
   <Contact />
 </div>;